Uses a separate remote and thumbnail cache
format
