+++ mobile/lib/modules/asset_viewer/image_providers/cache/image_loader.dart
