+++ mobile/lib/modules/asset_viewer/image_providers/cache/thumbnail_image_cache_manager.dart
@@ -1,0 +1,21 @@
+import 'package:flutter_cache_manager/flutter_cache_manager.dart';
+
+/// The cache manager for full size images [ImmichRemoteThumbnailProvider]
+class ThumbnailImageCacheManager extends CacheManager with ImageCacheManager {
+  static const key = 'remoteImageCacheKey';
+  static final ThumbnailImageCacheManager _instance =
+      ThumbnailImageCacheManager._();
+
+  factory ThumbnailImageCacheManager() {
+    return _instance;
+  }
+
+  ThumbnailImageCacheManager._()
+      : super(
+          Config(
+            key,
+            maxNrOfCacheObjects: 5000,
+            stalePeriod: const Duration(days: 30),
+          ),
+        );
+}
