+++ mobile/lib/modules/asset_viewer/image_providers/immich_remote_thumbnail_provider.dart
@@ -3,7 +3,8 @@ import 'dart:ui' as ui;
 
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter_cache_manager/flutter_cache_manager.dart';
-import 'package:immich_mobile/modules/asset_viewer/image_providers/image_loader.dart';
+import 'package:immich_mobile/modules/asset_viewer/image_providers/cache/image_loader.dart';
+import 'package:immich_mobile/modules/asset_viewer/image_providers/cache/thumbnail_image_cache_manager.dart';
 import 'package:openapi/api.dart' as api;
 
 import 'package:flutter/foundation.dart';
@@ -45,7 +46,7 @@ class ImmichRemoteThumbnailProvider
     ImageDecoderCallback decode,
   ) {
     final chunkEvents = StreamController<ImageChunkEvent>();
-    final cache = cacheManager ?? DefaultCacheManager();
+    final cache = cacheManager ?? ThumbnailImageCacheManager();
     return MultiImageStreamCompleter(
       codec: _codec(key, cache, decode, chunkEvents),
       scale: 1.0,
